# docs/openapi.json
{
	"openapi": "3.1.0",
	"info": { "title": "SocialAnimal REST", "version": "1.0.0" },
	"components": {
		"securitySchemes": {
			"bearerAuth": { "type": "http", "scheme": "bearer" }
		}
	},
	"paths": {
...
		"/api/v1/calendars": {
			"get": {
				"operationId": "calendars",
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 500,
							"default": 100
						}
					},
					{
						"name": "cursor",
						"in": "query",
						"required": false,
						"schema": { "type": "string" }
					}
				],
				"security": [{ "bearerAuth": [] }],
				"responses": {
					"200": {
						"description": "200",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"items": {
											"type": "array",
											"items": {
												"type": "object",
												"additionalProperties": false,
												"properties": {
													"id": { "type": "string" },
													"name": {
														"type": "string"
													},
													"type": {
														"type": "string"
													},
													"connectionId": {
														"type": [
															"null",
															"string"
														]
													},
													"remoteId": {
														"type": [
															"null",
															"string"
														]
													},
													"syncInterval": {
														"type": "integer"
													},
													"eventCount": {
														"type": "integer"
													},
													"lastSync": {
														"type": [
															"null",
															"string"
														]
													},
													"lastAttempt": {
														"type": [
															"null",
															"string"
														]
													},
													"lastSuccess": {
														"type": [
															"null",
															"string"
														]
													},
													"lastError": {
														"type": [
															"null",
															"string"
														]
													}
												},
												"required": [
													"id",
													"name",
													"type",
													"syncInterval"
												]
											}
										},
										"nextCursor": {
											"type": ["null", "string"]
										}
									},
									"required": ["items", "nextCursor"]
								}
							}
						}
					},
					"400": {
						"description": "400",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"401": {
						"description": "401",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"403": {
						"description": "403",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"404": {
						"description": "404",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"409": {
						"description": "409",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"422": {
						"description": "422",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"500": {
						"description": "500",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"502": {
						"description": "502",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"503": {
						"description": "503",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/calendars/{id}": {
			"patch": {
				"operationId": "updateCalendar",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"minLength": 1,
							"maxLength": 2048
						}
					}
				],
				"security": [{ "bearerAuth": [] }],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"additionalProperties": false,
								"properties": {
									"name": {
										"type": "string",
										"minLength": 1,
										"maxLength": 200
									},
									"syncInterval": {
										"type": "integer",
										"minimum": 0,
										"maximum": 525600
									}
								},
								"required": []
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "200",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"id": { "type": "string" },
										"name": { "type": "string" },
										"type": { "type": "string" },
										"connectionId": {
											"type": ["null", "string"]
										},
										"remoteId": {
											"type": ["null", "string"]
										},
										"syncInterval": { "type": "integer" },
										"eventCount": { "type": "integer" },
										"lastSync": {
											"type": ["null", "string"]
										},
										"lastAttempt": {
											"type": ["null", "string"]
										},
										"lastSuccess": {
											"type": ["null", "string"]
										},
										"lastError": {
											"type": ["null", "string"]
										}
									},
									"required": [
										"id",
										"name",
										"type",
										"syncInterval"
									]
								}
							}
						}
					},
					"400": {
						"description": "400",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"401": {
						"description": "401",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"403": {
						"description": "403",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"404": {
						"description": "404",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"409": {
						"description": "409",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"422": {
						"description": "422",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"500": {
						"description": "500",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"502": {
						"description": "502",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"503": {
						"description": "503",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "deleteCalendar",
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"minLength": 1,
							"maxLength": 2048
						}
					}
				],
				"security": [{ "bearerAuth": [] }],
				"responses": {
					"204": { "description": "204" },
					"400": {
						"description": "400",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"401": {
						"description": "401",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"403": {
						"description": "403",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"404": {
						"description": "404",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"409": {
						"description": "409",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"422": {
						"description": "422",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"500": {
						"description": "500",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"502": {
						"description": "502",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"503": {
						"description": "503",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					}
				}
			}
		},
		"/api/v1/events": {
			"get": {
				"operationId": "events",
				"parameters": [
					{
						"name": "limit",
						"in": "query",
						"required": false,
						"schema": {
							"type": "integer",
							"minimum": 1,
							"maximum": 500,
							"default": 100
						}
					},
					{
						"name": "cursor",
						"in": "query",
						"required": false,
						"schema": { "type": "string" }
					},
					{
						"name": "start",
						"in": "query",
						"required": true,
						"schema": { "type": "string", "format": "date-time" }
					},
					{
						"name": "end",
						"in": "query",
						"required": true,
						"schema": { "type": "string", "format": "date-time" }
					},
					{
						"name": "calendarId",
						"in": "query",
						"required": false,
						"schema": { "type": "string" }
					},
					{
						"name": "scope",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"enum": ["mine", "shared", "all"],
							"default": "all"
						}
					}
				],
				"security": [{ "bearerAuth": [] }],
				"responses": {
					"200": {
						"description": "200",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"items": {
											"type": "array",
											"items": {
												"type": "object",
												"additionalProperties": false,
												"properties": {
													"id": { "type": "string" },
													"calendarId": {
														"type": "string"
													},
													"title": {
														"type": "string"
													},
													"description": {
														"type": [
															"null",
															"string"
														]
													},
													"location": {
														"type": [
															"null",
															"string"
														]
													},
													"startTime": {
														"type": "string",
														"format": "date-time"
													},
													"endTime": {
														"type": "string",
														"format": "date-time"
													},
													"allDay": {
														"type": "boolean"
													},
													"isFriend": {
														"type": "boolean"
													},
													"calendar": {
														"type": "object",
														"additionalProperties": false,
														"properties": {
															"id": {
																"type": "string"
															},
															"name": {
																"type": "string"
															},
															"type": {
																"type": "string"
															}
														},
														"required": [
															"id",
															"name",
															"type"
														]
													},
													"owner": {
														"type": "object",
														"additionalProperties": false,
														"properties": {
															"id": {
																"type": "string"
															},
															"email": {
																"type": "string"
															},
															"name": {
																"type": [
																	"null",
																	"string"
																]
															}
														},
														"required": [
															"id",
															"email",
															"name"
														]
													}
												},
												"required": [
													"id",
													"calendarId",
													"title",
													"description",
													"location",
													"startTime",
													"endTime",
													"allDay",
													"isFriend",
													"calendar",
													"owner"
												]
											}
										},
										"nextCursor": {
											"type": ["null", "string"]
										}
									},
									"required": ["items", "nextCursor"]
								}
							}
						}
					},
					"400": {
						"description": "400",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"401": {
						"description": "401",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"403": {
						"description": "403",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"404": {
						"description": "404",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"409": {
						"description": "409",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"422": {
						"description": "422",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"500": {
						"description": "500",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"502": {
						"description": "502",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					},
					"503": {
						"description": "503",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"additionalProperties": false,
									"properties": {
										"code": { "type": "string" },
										"message": { "type": "string" },
										"requestId": { "type": "string" }
									},
									"required": ["code", "message", "requestId"]
								}
							}
						}
					}
				}
			}
		},
...
	}
}